fix: add dependencies to build ecdsa on mac (#1611)
## Description

Fix `make devenv-up` build failure for the `emily-cron` service on macOS
(specifically ARM64 architectures). The build previously failed during
the `uv sync` step when attempting to compile the `fastecdsa`
dependency, as necessary build tools (`gcc`) and development libraries
(`libgmp3-dev`) were missing from the base image. On x64, it was working
out-of-the-box because pypi offers the compiled binaries, and it was
skipping the compilation phase

Tested on @fabergat mac

## Changes

- Modified `docker/sbtc/emily-cron/Dockerfile`:
- Added `build-essential` and `libgmp-dev` to the `apt-get install`
command to ensure necessary build dependencies are available for
compiling Python packages with C extensions like `fastecdsa`.

## Testing Information

## Checklist:

- [ ] I have performed a self-review of my code
- [ ] My changes generate no new warnings
- [ ] New and existing unit tests pass locally with my changes
- [ ] Any dependent changes have been merged and published in downstream
modules