fix: add compound primary key to `rotate_keys_transactions` table (#1640)
## Description

This was supposed to be done in
https://github.com/stacks-sbtc/sbtc/pull/1599, but was missed. The bug
would make possible for us to drop key rotation events when they get
replayed following a fork.

## Changes

* Update the primary key of the `rotate_keys_transactions` table to be a
compound primary key on the `txid` and `block_hash` table.

## Testing Information

This PR adds an integration test to check that we can insert two events
with the same transaction ID into the database.

## Checklist:

- [x] I have performed a self-review of my code