feat: make the `getblock` RPC return detailed transacation info (#1642)
## Description

This is part of https://github.com/stacks-sbtc/sbtc/issues/1605, but
does not address it.

This PR makes some `BitcoinTxInfo` fields optional since coinbase
transactions omit them; this way we can reuse the same struct for both
`getblock` and `getrawtransaction` responses. We also dropped rarely
used fields to reduce memory overhead when processing blocks.

## Changes

* Remove several fields from the `BitcoinTxInfo` struct. It now only
contains what is necessary, except for the `fee` field which we keep for
convenience.
* Make some fields in the `BitcoinTxInfo` optional. Those fields were
never guaranteed to be present, so could have been omitted by bitcoin
core.
* Add a function that validates that we have the optional, but
required-for-us fields, in `BitcoinTxInfo`.
* Update the `getblock` command to return a detailed transaction objects
for a block.


## Testing Information

This is largely a refactor, but we should do some testing on private
mainnet before shipping this. According to this ticket
(https://github.com/bitcoin/bitcoin/issues/30495) we'll take a ~30x
performance slowdown compared to the status quo. And we want mainnet
testing because we have full blocks there.

## Checklist:

- [x] I have performed a self-review of my code