test: add devenv forking test for deposit (#1585)
## Description

Some work for [#1304](https://github.com/stacks-sbtc/sbtc/issues/1304),
[384](https://github.com/stacks-sbtc/sbtc/issues/384)

## Changes

Add a test for a forking scenario orphaning a deposit in devenv. The
annoying part is that it requires devenv running and with a bit of setup
before it can be run. And also it fails most of the times, but it's
ignored anyway.