fix: sign stacks tx once per tenure per request (#1691)
## Description

Closes: https://github.com/stacks-sbtc/sbtc/issues/1659

## Changes

Added (yet another) LRU to tx signers keeping the stacks "requests"
signed for each bitcoin tenure -- I went with a max size of 2, probably
1 would be enough as well. Also I went with the in memory solution
instead of the db since it's not something we need persistence for, as
we already only invoke the validation if the message comes from the
current coordinator.

On stacks tx signature request validation, we first check if it's known.
And after the signature broadcast, we add mark the request as signed.

~This covers only deposits and withdrawals, we could also add it for key
rotations but since it was not as susceptible to the issue as the others
I skipped it.~

## Testing Information

Added an integration test.

## Checklist:

- [ ] I have performed a self-review of my code
- [ ] My changes generate no new warnings
- [ ] New and existing unit tests pass locally with my changes
- [ ] Any dependent changes have been merged and published in downstream
modules