feat: enhance signers script pubkey check (#1722)
## Description

This came up as part of testing out the changes in related to adding a
new signer. Basically, a new signer should be able to validate a
complete-deposit contract call, so long as the sbtc signer backfilled
their bitcoin blockchain data with knowledge of the bootstrap aggregate
key.

## Changes

1. Update the `is_signer_script_pub_key` query to use information in the
`bitcoin_tx_outputs` outputs table in addition to the `dkg_shares`
table.

## Testing Information

This PR adds tests for the query and on the stacks validation logic,
which was affected by this change.

## Checklist:

- [x] I have performed a self-review of my code