feat: generate recoverable private shares during dkg (#1746)
## Description

Closes https://github.com/stacks-sbtc/sbtc/issues/1303

## Changes

* Generate secret shares using a random number generator that was seeded
with the signer's private key and the bitcoin block hash and height at
the time that the DKG started. This makes the secret shares
deterministically generated but still pseudo-random to outside
observers.
* Made a small refactor since the random number generator used when
processing messages is managed by the wrapping state machine.

## Testing Information

I added an integration test that checks that the DKG generates the same
secret shares when given the same inputs, but changes with a new bitcoin
block. I also added a test that makes sure that the nonces generated
during signing are still random.

## Checklist:

- [x] I have performed a self-review of my code