feat: forgo unnecessary db query when fetching pending deposits (#1774)
## Description

I just noticed that we made an unnecessary database call. It happens at
most once per bitcoin block, so no big deal, but figured we should do
without it anyway.

## Changes

* Changed the type signature of `get_pending_accepted_deposit_requests`
to take a `BitcoinBlockRef` instead of a `BitcoinBlockHash` so that we
can forgo the database call.

## Testing Information

I'm not sure what tests should be added. If I think of something, I'll
add it.

## Checklist

- [x] I have performed a self-review of my code