feat: set the sign and dkg IDs in wsts coordinator state machines (#1776)
## Description

Closes https://github.com/stacks-sbtc/sbtc/issues/1177 and closes
https://github.com/stacks-sbtc/sbtc/issues/1775.

## Changes

* Set the `dkg_id` in the WSTS state machines. The `dkg_id` is set to
the bitcoin block height associated with the DKG round that generated
the keys.
* Set the `sign_id` during all signing rounds. It is set to the first 8
bytes of the SHA-256 of the message and the bitcoin chain tip at the
time that the signing round took place.

These values are only set in the coordinator state machine and are
inherited in the signer state machine. This ends up being okay, since
out checks in the tx-signer is validation that we would do in the state
machine.

### Notes

We do not set the `sign_id` in the FROST coordinator because that field
is private there. This is okay, because I want to remove our use of the
FROST coordinator (and remove it generally
https://github.com/stacks-sbtc/wsts/issues/180), which we can do once
https://github.com/stacks-sbtc/wsts/issues/174 is done.

Also, WSTS internally changes the sign ID for you, which is kinda
helpful but not at the same time. For us that means that the expected
sign ID is always off by one.

## Testing Information

I added an integration test that checks that the values send over the
wire have the expected `dkg_id` and `sign_id`.

## Checklist

- [x] I have performed a self-review of my code