chore: fix the `bitcoin-miner`'s script in `devenv` to not crash if bitcoin is unavailable (#1794)
## Description

Fixes the devenv `bitcoin-miner` service's entrypoint script to
gracefully handle "hiccups" in communication with bitcoin-core.

Previously if you stopped `bitcoind`, the `bitcoin-miner` container
would crash and you'd be unable to start it again due to unloaded
wallets, etc.

This was a prerequisite to properly testing the
[ZMQ->poller](https://github.com/stacks-sbtc/sbtc/pull/1780) changes in
`devenv`.

Note that this does allow `bitcoin-miner` to also be stopped/started,
but at present it'll still run the `generatetoaddress` calls resulting
in ~200 extra blocks the signers need to sift through. This PR _doesn't_
address this.

## Changes

- Refactor the logic in the `bitcoin-miner` mining script.
- Move the script into its own file to be able to get IDE
support/highlighting and reduce bloat of the `docker-compose.yml` file.

## Testing Information

Tested thoroughly in devenv. Note that you must **stop** the `bitcoind`
container, not `down` it, otherwise you'll lose your chainstate.

## Checklist

- [x] I have performed a self-review of my code
- [x] My changes generate no new warnings
- [x] New and existing unit tests pass locally with my changes
- [x] Any dependent changes have been merged and published in downstream
modules

---------

Co-authored-by: Matteo Almanza <matteo.almanza@algorand.com>