feat: relaxed handling of v2 pox responses (specifically epochs) (#1836)
## Description

Closes: #1816 

## Changes

- Adds a new `StacksInteract::get_epoch_status()` method which uses a
minimal struct for deserialization of `/v2/pox` responses, treating
`epoch_id` as a string instead of `StacksEpochId`. This method returns a
new `StacksEpochStatus` enum variant which covers the two cases that
we're interested in: "are we in 3.0+, or not?", and associated metadata
which is used for logging/logic elsewhere (nakamoto start height and
current height).
- Replaced the usage of `get_pox_info()` with the new method everywhere
(including tests).
- Added additional tests specifically testing the epoch 3.0 boundary,
including new/garbage epochs which we're not interested in (at the
moment).


## Testing Information

All tests pass and devenv seems to work fine.

## Checklist

- [x] I have performed a self-review of my code
- [x] My changes generate no new warnings
- [x] New and existing unit tests pass locally with my changes
- [x] Any dependent changes have been merged and published in downstream
modules