fix: handle testnet stacks blocks backfill (#1850)
## Description

Closes: https://github.com/stacks-sbtc/sbtc/issues/1749
Supersedes: https://github.com/stacks-sbtc/sbtc/pull/1803

## Changes

During Stacks blocks backfill, if we encounter a 404 when fetching the
tenure for a block we also check if that block is a pre-Nakamoto one
and, if so, we stop the backfilling process.

## Testing Information

Added a green ignored integration test targeting testnet
(`fetch_unknown_ancestors_works_in_testnet`); devenv backfill worked as
expected.

## Checklist

- [ ] I have performed a self-review of my code
- [ ] My changes generate no new warnings
- [ ] New and existing unit tests pass locally with my changes
- [ ] Any dependent changes have been merged and published in downstream
modules