Add Emily Sidecar And Cron Pipelines (#1843)
I've added emily sidecar and cron pipelines in order to deploy to ecr,
using the same method of running the CI as for the emily deployment. The
core logic here is built by @fabergat in
https://github.com/stacks-sbtc/sbtc/pull/1830, on top of which I've
brought the modifications to the way in which the workflow is run made
in https://github.com/stacks-sbtc/sbtc/pull/1840.

---------

Co-authored-by: Fabrizio Gattuso <f.gattuso87@gmail.com>