[chore]: add machete (#1852)
## Description

Closes: no issue

## Changes

Adds `cargo-machete`, a tool for detecting unused dependencies in
`Cargo.toml`.

Note: this tool does not give a 100% guarantee that it will find _all_
unused dependencies, so we still should look if the dependency is
necessary during a review.
Note 2: There is a similar tool named `cargo-udeps`, so maybe we should
consider using it/both.

## Testing Information
no testing needed
## Checklist

- [ ] I have performed a self-review of my code
- [ ] My changes generate no new warnings
- [ ] New and existing unit tests pass locally with my changes
- [ ] Any dependent changes have been merged and published in downstream
modules