fix: release builds broken because of carelessly hidden impls (#1856)
## Description

## Changes

Restores `impl`s that were marked as testing in
https://github.com/stacks-sbtc/sbtc/pull/1851.

## Testing Information

I tested this with `cargo build --bin signer --release --locked`, which
is the same command we use in our Dockerfile for release builds. It
would be nice if CI would catch this kind of thing. Basically, our usual
CI could be fine for debug builds but broken for release builds.

## Checklist

- [x] I have performed a self-review of my code