chore: bump version of stacks core dependencies to 3.3.0.0.1 (#1871)
## Description

Stacks-core introduced some new variants of the `TenureChangeCause` enum
in the latest upgrade, and that enum is used to construct Stacks
transactions. Since we use stacks-core types directly when deserializing
Stacks blocks, signers won't be able to process Stacks blocks that
include transactions that make use of the new enum variants. So, if such
transactions were to be included in a block, we would fail to decode
them.

## Changes

1. Bump the version of our stacks core dependency crates to the commit
of 3.3.0.0.1

## Testing Information

Tested deposits and withdrawals on devenv and everything worked
smoothly.

## Checklist

- [x] I have performed a self-review of my code