feat: improve stacks blocks backfilling (#1857)
## Description

Closes https://github.com/stacks-sbtc/sbtc/issues/1817.

## Changes

1. Adds a new table for writing stacks blocks to "temporarily".
2. Adds new functions for writing to the new temporary table, truncating
it, and copy its contents to the `stacks_blocks` table.
3. Modifies the `fetch_unknown_ancestors` function to write blocks to
the database one tenure's worth of blocks at a time. Also renames the
function so that it better aligns with what it does now.
4. Marked some trait methods as for testing only, since that's where
they are currently used.
5. I changed the `cargo build` command used in `make build` to use
`--bins --lib` instead of `--all-targets`. Using `--all-targets` meant
that test binaries were built -- according to
https://doc.rust-lang.org/cargo/commands/cargo-build.html -- which meant
that the `testing` feature in the signer crate was enabled in both the
testing binary and the main binary due to how cargo unifies features
when building targets. However, we do not want this so I included only
the necessary targets when compiling a debug and release builds.

I added the change in (5) here because I used it to make sure that
things were built correctly for my devenv tests, but I can see it
warrenting it's own PR. So if anyone has any strong opinions I can PR
that commit too.

## Testing Information

I tested everything up to commit
4bed294a82c402a7213e8e57b739c730a43e86de on devenv. Everything worked
fine. I also made sure that the `--all-targets`/`--bins --lib` switch
worked as expected. I also build a release image to make sure that
everything still at least build correctly and that succeeded.

## Checklist

- [x] I have performed a self-review of my code