feat: add dynamic context to Emily (#1867)
## Description

Work for https://github.com/stacks-sbtc/sbtc/issues/1854

## Changes

Pass the Emily context (table names and dynamo db client) dynamically to
the handlers, instead of passing it at startup time; since it's dynamic
we can now change its value for tests -- in particular we can (in some
future PR) change the table names to run the tests against a single
`emily-server` instance.

This PR also add a flag to `emily-server` to avoid looking for table
names at startup, it's not used right now but will be required to use a
single dynamodb instance in parallel tests.

## Testing Information

Add a test against the `/health` endpoint to test changing the version.

## Checklist

- [X] I have performed a self-review of my code
- [X] My changes generate no new warnings
- [X] New and existing unit tests pass locally with my changes
- [X] Any dependent changes have been merged and published in downstream
modules